Revert "1.6.2. Тоже отклинапенный."
This reverts commit 96ac53bc30e53bc7eeae05d14faa6636619e9783.

+++ Minecraft/src/net/minecraft/src/IEntityMultiPart.java
@@ -2,7 +2,7 @@ package net.minecraft.src;
 
 public interface IEntityMultiPart
 {
-	boolean attackEntityFromPart(EntityDragonPart var1, DamageSource var2, float var3);
+	boolean attackEntityFromPart(EntityDragonPart var1, DamageSource var2, int var3);
 	
 	World func_82194_d();
 }
